set better defaults

+++ src/pdfflow/configflow.py
@@ -4,8 +4,8 @@ Define some constants, header style
 # Most of this can be moved to a yaml file without loss of generality
 import tensorflow as tf
 
-# Define the tensorflow numberic types
-DTYPE = tf.float32
+# Define the tensorflow number types
+DTYPE = tf.float64
 DTYPEINT = tf.int32
 
 # Create wrappers in order to have numbers of the correct type
